Moved CheckersWebSocketHandler to separate package for socket handlers
