+++ backend/src/main/java/pw/checkers/config/WebSocketConfig.java
@@ -3,6 +3,7 @@ package pw.checkers.config;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.socket.config.annotation.*;
 import pw.checkers.service.GameService;
+import pw.checkers.sockets.CheckersWebSocketHandler;
 
 @Configuration
 @EnableWebSocket
