+++ backend/src/main/resources/application-dev.properties
@@ -1,1 +1,1 @@
-logging.level.pw.checkers.config.CheckersWebSocketHandler=DEBUG
+logging.level.pw.checkers.sockets.CheckersWebSocketHandler=DEBUG
